testing done with coverage 82% of lines

+++ Backend/src/main/java/com/gameshub/model/request/DigitalProductRequestDAO.java
@@ -14,7 +14,7 @@ public class DigitalProductRequestDAO extends ProductRequestDAO {
     @Column(name = "Code")
     private String code;
 
-    public DigitalProductRequestDAO(LocalDate dateReceived, String status, String requestType, String title, int price, String description, LocalDate postDate, int count, String category, SellerDAO seller, String code) {
+    public DigitalProductRequestDAO(LocalDate dateReceived, String status, String requestType, String title, float price, String description, LocalDate postDate, int count, String category, SellerDAO seller, String code) {
         this.dateReceived = dateReceived;
         this.status = status;
         this.requestType = requestType;
